+++ LookAroundEyesApp/src/Math.h
@@ -61,7 +61,7 @@ struct vec3f
         return (lhs.x * rhs.x) + (lhs.y * rhs.y) + (lhs.z * rhs.z);
     }
     static float length(const vec3f& v) {
-        return sqrtf(v.x * v.x + v.y * v.y + v.z * v.z);
+        return static_cast<float>(std::sqrt(v.x * v.x + v.y * v.y + v.z * v.z));
     }
     static vec3f normalize(const vec3f& v) {
         float len = length(v);
@@ -133,8 +133,8 @@ struct mat4f
     // Rotation around Y axis
     static mat4f rotationY(float angle) {
         mat4f result = identity();
-        float c = cosf(angle);
-        float s = sinf(angle);
+        float c = static_cast<float>(std::cos(angle));
+        float s = static_cast<float>(std::sin(angle));
         result.m[0][0] = c;
         result.m[0][2] = s;
         result.m[2][0] = -s;
@@ -145,8 +145,8 @@ struct mat4f
     // Rotation around X axis
     static mat4f rotationX(float angle) {
         mat4f result = identity();
-        float c = cosf(angle);
-        float s = sinf(angle);
+        float c = static_cast<float>(std::cos(angle));
+        float s = static_cast<float>(std::sin(angle));
         result.m[1][1] = c;
         result.m[1][2] = -s;
         result.m[2][1] = s;
@@ -157,8 +157,8 @@ struct mat4f
     // Rotation around Z axis
     static mat4f rotationZ(float angle) {
         mat4f result = identity();
-        float c = cosf(angle);
-        float s = sinf(angle);
+        float c = static_cast<float>(std::cos(angle));
+        float s = static_cast<float>(std::sin(angle));
         result.m[0][0] = c;
         result.m[0][1] = -s;
         result.m[1][0] = s;
